requirements.txt
streamlit
fastapi
uvicorn
sympy
python-multipart
requests
transformers
torch
torchvision
pillow
numpy
pandas
scikit-learn
tqdm
latex
